+++ Payroll_Project2/Forms/Employee/File leave/fileLeaveUC.cs
@@ -31,7 +31,7 @@ namespace Payroll_Project2.Forms.Employee.File_leave
         private DateTime EndingDate { get; set; }
         private string LeaveDetails { get; set; }
         private int NumberOfDays { get; set; }
-        private float CreditsUsed { get; set; }
+        private decimal CreditsUsed { get; set; }
 
 
         public fileLeaveUC()
@@ -42,7 +42,7 @@ namespace Payroll_Project2.Forms.Employee.File_leave
         #region Functions responsible for communicating with the general functions class
 
         private async Task<bool> AddleaveRequest(int applicationNumber, int employeeId, DateTime dateFile, string leaveType, string formType,
-            string leaveDetails, int numberOfDays, DateTime leaveStartDate, DateTime leaveEndDate, float creditsUsed, string status)
+            string leaveDetails, int numberOfDays, DateTime leaveStartDate, DateTime leaveEndDate, decimal creditsUsed, string status)
         {
             try
             {
@@ -100,11 +100,11 @@ namespace Payroll_Project2.Forms.Employee.File_leave
             catch (Exception ex) { throw ex; }
         }
 
-        private async Task<float> GetLeaveCredits(int employeeId, string leaveType, int year)
+        private async Task<decimal> GetLeaveCredits(int employeeId, string leaveType, int year)
         {
             try
             {
-                float getLeaveCredits = await generalFunctions.GetLeaveCredits(employeeId, leaveType, year);
+                decimal getLeaveCredits = await generalFunctions.GetLeaveCredits(employeeId, leaveType, year);
 
                 return getLeaveCredits;
             }
@@ -414,11 +414,11 @@ namespace Payroll_Project2.Forms.Employee.File_leave
             }
         }
 
-        private async Task<bool> IsAllowed(int employeeId, string leaveType, int year, float numberOfCredits)
+        private async Task<bool> IsAllowed(int employeeId, string leaveType, int year, decimal numberOfCredits)
         {
             try
             {
-                float getCredits = await GetLeaveCredits(employeeId, leaveType, year);
+                decimal getCredits = await GetLeaveCredits(employeeId, leaveType, year);
 
                 if (getCredits > 0 && numberOfCredits <= getCredits)
                 {
