+++ Payroll_Project2/Forms/Personnel/Personal Portal/File Leave/fileLeaveUC.cs
@@ -31,7 +31,7 @@ namespace Payroll_Project2.Forms.Personnel.Personal_Portal.File_Leave
         private DateTime EndingDate { get; set; }
         private string LeaveDetails { get; set; }
         private int NumberOfDays { get; set; }
-        private float CreditsUsed { get; set; }
+        private decimal CreditsUsed { get; set; }
 
 
         public fileLeaveUC(int userId, newDashboard parent)
@@ -44,7 +44,7 @@ namespace Payroll_Project2.Forms.Personnel.Personal_Portal.File_Leave
         #region Functions responsible for communicating with the general functions class
 
         private async Task<bool> AddleaveRequest(int applicationNumber, int employeeId, DateTime dateFile, string leaveType, string formType, 
-            string leaveDetails,int numberOfDays, DateTime leaveStartDate, DateTime leaveEndDate, float creditsUsed, string status)
+            string leaveDetails,int numberOfDays, DateTime leaveStartDate, DateTime leaveEndDate, decimal creditsUsed, string status)
         {
             try
             {
@@ -99,11 +99,11 @@ namespace Payroll_Project2.Forms.Personnel.Personal_Portal.File_Leave
             catch (SqlException sql) { throw sql; } catch (Exception ex) { throw ex; }
         }
 
-        private async Task<float> GetLeaveCredits(int employeeId,  string leaveType, int year)
+        private async Task<decimal> GetLeaveCredits(int employeeId,  string leaveType, int year)
         {
             try
             {
-                float getLeaveCredits = await generalFunctions.GetLeaveCredits(employeeId, leaveType, year);
+                decimal getLeaveCredits = await generalFunctions.GetLeaveCredits(employeeId, leaveType, year);
 
                 return getLeaveCredits;
             }
@@ -409,11 +409,11 @@ namespace Payroll_Project2.Forms.Personnel.Personal_Portal.File_Leave
             }
         }
 
-        private async Task<bool> IsAllowed(int employeeId, string leaveType, int year, float numberOfCredits)
+        private async Task<bool> IsAllowed(int employeeId, string leaveType, int year, decimal numberOfCredits)
         {
             try
             {
-                float getCredits = await GetLeaveCredits(employeeId, leaveType, year);
+                decimal getCredits = await GetLeaveCredits(employeeId, leaveType, year);
 
                 if(getCredits > 0 && numberOfCredits <= getCredits)
                 {
